requirements.txt
redis==5.0.0
psycopg2-binary==2.9.9
streamlit==1.30.0
pandas==2.1.2
numpy==1.26.4
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
streamlit-autorefresh==1.0.1
